requirement.txt
asgiref==3.5.2
attr==0.3.2
attrs==22.1.0
cached-property==1.5.2
certifi==2022.9.14
charset-normalizer==2.1.1
contourpy==1.0.5
cycler==0.11.0
Django==4.0.6
django-isbn-field==0.5.3
fonttools==4.37.2
httplib2==0.20.4
idna==3.4
isodate==0.6.1
kiwisolver==1.4.4
lxml==4.9.1
matplotlib==3.6.0
numpy==1.23.3
packaging==21.3
Pillow==9.2.0
platformdirs==2.5.2
pycairo==1.21.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-stdnum==1.17
pytz==2022.2.1
requests==2.28.1
requests-file==1.5.1
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.12
zeep==4.1.0
